backend/requirements.txt
Django==3.2.20
django-colorfield==0.9.0
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
gunicorn==20.1.0
isort==5.10.1
Pillow==9.1.1
psycopg2-binary==2.9.7
python-dotenv==0.20.0Ы